Create a container for the TLV data starting from the 4th element
